+++ Makefile
@@ -16,9 +16,8 @@ LIBFT = $(DIR_LIBFT)libft.a
 NAME = ./minishell
 SRC = ./src/
 INCLUDE = ./include
-FILES = btree_util_1.c \
-		btree_util_2.c \
-		cd.c \
+FILES = cd_1.c \
+		cd_2.c \
 		echo_1.c \
 		echo_2.c \
 		echo_3.c \
@@ -27,15 +26,15 @@ FILES = btree_util_1.c \
 		export_1.c \
 		export_2.c \
 		export_3.c \
+		lstnew_util_1.c \
+		lstnew_util_2.c \
 		main.c \
 		other_command.c \
 		pwd.c \
 		split_2.c \
 		unset.c \
 		util_1.c \
-		util_2.c \
-		lstnew_util_1.c \
-		lstnew_util_2.c
+		util_2.c
 
 
 SRC_FILES = $(addprefix $(SRC),$(FILES))
